refactor: migrate new functions to numeric.h
